requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.7
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
python-twitter==3.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
tablib==3.0.0
twitter==1.18.0
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
